Adding tests to the data module. Some basic tests are written, but can be made different. TBD

+++ domain/src/main/java/com/helpfulapps/domain/repository/AlarmRepository.kt
@@ -8,9 +8,9 @@ interface AlarmRepository {
 
     fun getAlarms(): Single<List<Alarm>>
 
-    fun removeAlarm(alarmId: Int): Completable
+    fun removeAlarm(alarmId: Long): Completable
 
-    fun switchAlarm(alarmId: Int): Completable
+    fun switchAlarm(alarmId: Long): Completable
 
     fun addAlarm(alarm: Alarm): Completable
