+++ domain/src/main/java/com/helpfulapps/domain/use_cases/alarm/SwitchAlarmUseCase.kt
@@ -5,7 +5,7 @@ import com.helpfulapps.domain.use_cases.type.CompletableUseCaseWithParameter
 import io.reactivex.Completable
 
 class SwitchAlarmUseCase(private val repository: AlarmRepository) :
-    CompletableUseCaseWithParameter<Int> {
+    CompletableUseCaseWithParameter<Long> {
 
-    override fun execute(alarmId: Int): Completable = repository.switchAlarm(alarmId)
+    override fun execute(alarmId: Long): Completable = repository.switchAlarm(alarmId)
 }
